parking_management/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.11.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.2.0
crispy-bootstrap4==2022.1
crypto==1.4.1
cryptography==44.0.2
cycler==0.12.1
distlib 
dj-rest-auth==7.0.1
Django==4.2.20
django-allauth==65.7.0
django-auditlog==3.0.0
django-bootstrap4==22.3
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-dynamic-datatb==1.0.25
django-filter==25.1
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
filelock 
fonttools==4.49.0
frozenlist==1.5.0
idna==3.10
importlib-resources==6.1.2
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.4.3
Naked==0.1.32
numpy==1.26.4
packaging==23.2
pandas==2.2.1
Pillow==9.4.0
platformdirs 
plotly==5.24.1
propcache==0.3.1
psycopg2==2.9.5
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2022.7.1
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
six==1.16.0
soupsieve==2.4
sqlparse==0.5.3
tenacity==9.0.0
twilio==9.5.2
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
virtualenv 
virtualenvwrapper-win 
yarl==1.19.0
zipp==3.17.0
